+++ src/main/java/com/musimundo/feeds/service/FeedBuilderService.java
@@ -1,0 +1,6 @@
+package com.musimundo.feeds.service;
+
+public interface FeedBuilderService {
+
+    void createRegister(String path);
+}