+++ src/main/java/com/musimundo/feeds/service/ProcessingFeedService.java
@@ -1,0 +1,7 @@
+package com.musimundo.feeds.service;
+
+import com.musimundo.utilities.FeedType;
+
+public interface ProcessingFeedService {
+    void process(FeedType feedType);
+}