+++ src/main/java/com/musimundo/utilities/ErrorType.java
@@ -1,0 +1,5 @@
+package com.musimundo.utilities;
+
+public enum ErrorType {
+    SUCCESS, ERROR, WARNING
+}